.emacs: switch to non-cycling histdir up/down
I originally found the cycling weird when I
switched from typical shells to Eshell, but I
stuck with it to see how I liked it, and it's
not terrible, but realistically, the main
difference it makes is that I semi-regularly
cycle around to the oldest history entries on
accident, and the main intentional use of
history cycling I have is to undo that.

Leaving the cycling functions in the code for
now because I'm thinking of this as a package,
and this is a useful behavior variation where
a user might want one or the other or both (I
myself want non-cycling down, non-cycling up,
and cycle-newer behind some other keybind to
see the oldest entries because that's a good
start for finding history entries which I
probably don't want to keep).