.emacs: combine lambda-let and lambda-partial-let
I considered doing this back when I first added
lambda-partial-let, but held off because it seemed
like I might have enough uses where it would be
nice to not have an argument list. But a lot of the
time when I'm fiddling, I end up needing to specify
an argument list anyway, and in some ways it's a
simpler concept to just have one: it's like lambda
combined with a let, and let takes a variable list
while lambda takes an argument list - so lambda-let
takes both.