.emacs: fix fixed-consult-history preview position with prefix argument
I basically never use this feature, but I've been
keeping it, except behind a prefix argument and
made sane, since it's in `consult-history`. Guess
I forgot to properly test this during the last
relevant round of changes (when restoring preview
capability and making it work with the histdir+vi
wrapped REPLs).