.emacs: make "gh"/"gr" for tags not change "title"
This is basically exclusively for preserving "--"
nested inside already-sluggified titles. I want
to stay within the Denote naming convention as
much as possible, but I really want some kind of
capability for stronger/weaker separation within
the file title (for example, when a song has two
official titles, or when there are two or more
concepts in the title and single-dash separation
both within and between them makes that harder
to read/etc).

However, the least-effort (and more flexible for
anything else I chose to do with titles that
Denote wouldn't do) solution is to just advise
the sluggify operation into a no-op in these
operations where the title is already a slug.