.emacs: make "gh"/"gr" for tags not change "tags"
This is primarily for consistency with how the
tags appear in the file name: if I have a tag
like "foo-bar", Denote keeps the `-` in the
file name, but deletes it in the frontmatter.
I would rather it be consistent, and I never
make use of the feature of stripping any other
characters from the tags, so this is probably
a safe-enough change.