.emacs: " m" consistent with "gh"/"gr" for tags
Keep the "sluggifying" from happening on note
creation since I don't use it when changing
tags on existing files either.

There *is* a little nicety in being able to
type "foo bar" and have it automatically turn
into "foo-bar", but I'd rather give that up
and manually type "-" for the benefit of not
needing to manually remove and then re-add
tags which have dashes in them.

In the longer term maybe I will add the logic
to replace " " with "-" in both title and tag
strings instead of having no/stock slugging.