.emacs: remove use-package -> no-op guard
Now that I'm using Emacs 29.1 on all of my devices,
I don't need this to guard against use-package
being entirely uninstalled. I might still want a
similar guard for when setting up Emacs with no
packages installed yet, but we'll cross that
bridge when we get there.

(I've thought for some time that I might want a
CLI flag to flip on the "ensure" option, for
example.)