.emacs: improve y-or-n-p instead of yes-or-no-p
Using an alias worked for some functions but not
others - in particular, it didn't work for
`kill-buffer`, which is a very common path which
basically never requires multiple keypresses of
interruption to think over, so it was starting
to get really annoying.

Not sure when upstream added `use-short-answers`
or if I missed it or had a reason for not using
it earlier, but it seemes to do the same thing
at a deeper level with only mild aesthetic
differences (different font colors).