.emacs: REPL wrapper: fix mouse click in insert and replace states
This is something I never use on my computers,
but on phones it ends up being naturally more
efficient to tap to move the cursor around in
some cases.

Been putting off dealing with this for a few
months, finally got some time to deal with it,
and the nagging->priority from tripping over
it accumulated a bit.

It turned out to be trivial, but for some
reason Replace state needs the down-mouse-1
event to be overridden because it weirdly
ignores the mouse-1 event. I've been using
mouse-1 instead of down-mouse-1 in every
other keybinding, so keeping Replace state
as the exception for now - no time to figure
this out any deeper at the moment.