.emacs: make " g" save+kill buffer and go to task in Dired
This makes the UX a lot like a task list (such as
Todoist and clones) where you get a text box in
which you type the task, hit enter, and it just
appears in the list itself.

(Do I actually want the Denote notes folder to
pop open? Only in the case where I want to add
tags or edit the note. But it only takes one
keypress ("q") to make it go away. Conversely,
" M" will basically take me to the same place.)