.emacs: " j" to delete and recreate a task
(Mnemonic: task jumps to the end of the task
queue, although I'm not super committed to this
keybind for it. Also, a "J" even looks like a
path which could be dropping down and then
looping back around to stop at the bottom.)

Useful when you've made progress on a task
but need to come back to it. For an example,
the "Autofocus" method heavily uses this, and
I am implementing something heavily inspired
by it.

GTD would probably recommend splitting up
tasks if this happens. (But even then, a file
tagged with my task tag doesn't have to be one
task, it can be a project note containing many
tasks.)

Also this provides a basic priority scheme:
push back low-priority items, high priority
bubbles up to the top. (And once you start
to recognize an item as regularly floating
up only to be sent back down, that's a good
cue to reconsider keeping it at all.)