.emacs: fix org-read-date-display error handling
Stock behavior breaks preview on the first error,
including parsing "errors" such as if you typo
"1:11" as "1:1q". After the first error, there
is no more preview even once input is correct.

This ensures a more sensible behavior of just
reporting the errors and removing the preview
but then restoring the preview as soon as the
input is valid again.