.emacs: org-read-display: decouple preview and calendar options
For some reason, `org-read-display` ignores the
preference for live preview when the calendar
is not displayed. I understand that there might
be some edge cases which I missed where it breaks
in the absence of a calendar. But given that it is
objectively bad design to complect two unrelated
behaviors together, I have elected to force it to
do the actually sensible behavior where these two
unrelated features are not complected together.