.emacs: make org-read-date not pop calendar
This would be a neat feature if the calendar
was fully-featured: showed all days with
scheduled events in distinct color, showed
(or had a keybind away) a view of the events
for the day you're on, and so on. Also if
the calendar wasn't idiotically hard-coded
to three months, or was at least easier to
monkeypatch (as is, it's a hardcoded 3 in a
`dotimes` form), since that works poorly on
mobile.