.emacs: rewrite Denote-based task, note, and file tag/title/timestamp logic
Not exactly a full rewrite, but lots of changes,
which I just don't have it in me to pull apart
into clean atomic commits. Maybe later, but
probably not. I will probably regret the loss of
history granularity later, but life is short and
I am tired.

The Denote-style file name manipulations are all
moved under the umbrella "denoted" (because the
file name parts are denoted! and because it's a
better name than denote-ish or denotesque}). This
include tags, title, and (newly added) timestamp.

Logic for Denoted file name manipulations has been
unified and simplified.

Denoted file renaming logic now supports replacing
the file "ID" (this doesn't affect the frontmatter,
due to a limitation/assumption in Denote which I
currently have no real reason to fix, since I am
not yet using Denote links and my foreseeable uses
for Denote links are for actual notes, not for task
files which just happen to be implemented as notes).
This uses a custom prompt which does minimal parsing
of the timestamp to get a Denote-style ID string.
This is bound on "gs" (originally the mnemonic was
for "schedule", but that's a stretch... it's one of
the remaining unbound "g"-prefixed keybindings, and
at this point it's more for consistency with the
others than anything else.)

Denoted file renaming logic now doesn't add
frontmatter if it wasn't already there (but does
update the frontmatter if it already exists).

Misc. minor changes:

* denote-title-prompt now never suggests nor
  remembers history (but a separate candidates
  variable is available if there's ever a good
  reason to title candidates through minibuffer
  completion).
* Task "recreate" is now based on the Denoted
  timestamp setting function instead of the
  earlier hack.
* `task-list` and the refresh after task create,
  "recreate", and scheduling now properly handle
  non-task files existing "after" (sorted later
  in the Dired buffer) the last task, the cursor
  still ends up on the task after filtering even
  in those cases.
* Task "schedule" still uses the Org prompt for
  getting the date+time. I don't like Org's prompt
  and I dislike Org's code and 0.5s loading time
  of its single 20k LoC source file, but this does
  allow me nicer UX, since I don't want to spend
  more time just yet on implementing my own vision
  for a better (for me at least) date+time parsing
  function. This prompt intentionally ignores the
  existing timestamp. This means that small
  adjustments to task time can be done with the
  Denoted timestamp edit keybind, which pulls up
  the existing timestamp, while free-form scheduling
  into the future can be more ergonomically done
  with Org's fancier prompt.

Plus a bunch of code code re-ordered for greater
cohesiveness (related things closer together,
plus more pattern in how they're ordered).

Might've missed something.