.emacs: datetime-parse: larger units default to now
If I specify "January", I presumably mean *this*
January, not the January in year 0. Generalized:
if a person specifies a high-precision time unit,
the natural impliciation is that it's meant in
the scope of whatever lower-precision time units
we are in right now.

On the other hand, if I specify "January", I
*don't* yet mean any specific day, hour, minute,
and so on.