.emacs: `setq-if-nil` macro
Finally wrapping up this pattern that
I use in so many places. Basically,

    (setq-if-nil foo baz)

replaces

    (unless foo
        (setq foo baz))

but also supports the `setq` convenience
of multiple assignments in one expression,
so

    (setq-if-nil foo baz qux alp)

replaces

    (unless foo
        (setq foo baz))
    (unless qux
        (setq qux alp))

and as a finishing touch of consistency
and sensibility in the edge cases, it
returns the value that the variable has
after it runs. So if the value was already
not nil, it returns that value, and if the
value was nil, it returns the new value.

I'm not 100% on the rightness of that final
decision - always returning the passed-in
value, even if there was a non-nil value
already, might prove more useful in the
situations where you'd actually want the
passthrough. We'll see. But for now, it
seems conceptually righter to return the
value that actually results from this
default-if-nil assignment.