.emacs: datetime-parse: reject "st" offset suffix
Generally tolerating junk characters after the
minimum unambiguous characters seems fine, but
"st" collides with the "st" suffix for days,
which is valid for inputs like "1st" but not
for inputs like "+1st" and "-1st".

Since "1st" gets a different interpretation than
"1s", it's proper to make sure that "+1st" gets
a different interpretation than "+1s".