.emacs: " ]" to toggle `select-enable-clipboard`
This solves a long-standing bug I occasionally
see with trying to copy large chunks of text on
Android in Termux.

The actual error seems to happen outside of Emacs,
because it leaves junk text in the terminal which
Emacs doesn't know it needs to redraw (evidenced
by how Emacs only redraws the parts of the screen
that Emacs already had text in).

I'm not even sure that it's a problem for Emacs,
but the UX is a problem - I can tolerate some
visual glitching and manual overhead to clear it,
but I can't tolerate a bunch of cues that seem
like the copy/cut failed and messed up the buffer.