.emacs: try " a" to bring up calendar
For now, anyway. Good odds I'll change it,
but I've been thinking for some time that
" a" would be for agenda-style things, so
this is a natural place to hold it.

I imagine I end up redefining a lot of keys
in the calendar. For example, instead of
diary entries, I could make the calendar
work with my task files (based on Denote).