.emacs: finish unbinding "H" as vi motion
For months I've been using H to do what stock
vi's J does, J and K to do what H and L do,
and H and L for misc. "back" and "forward"
keybinds. This makes a lot of sense to me,
(especially since J/K correspond directly to
j/k this way), and has been working great.

But apparently I had left "H" bound in motion
state, which I mostly haven't noticed since
it's shadowed by "H" rebound in normal state
(to hoist the next line up and join it to the
current line) and by my mode-specific "H"/"L"
bindings.

I don't remember if it was intentional or not
that I left it bound in motion state.