.emacs: new Orderless "!" suffix to negate
Unlike `orderless-without-literal`, my "!"
implementation seemlessly integrates with
my `just-backslashes` match style, so that

* "foo!" unmatches "foo",
* "foo\!" matches "foo!",
* "foo\\!" unmatches "foo\",

and so on.

I would have loved to negate a regex instead
of negating a literal, but that basically does
not exist (at least not with the features of
Emacs' regular expression engine, but even in
the most featureful engines it's a hack using
lookahead or possessive quantifiers or atomic
groups or whatever) and is very non-trivial to
do.

So instead, for consistency, I am disabling
regex match entirely (in prior commit).
Regular expressions are great and I do find
them very helpful at times. But in minibuffer
completion/narrowing/search, the only needs of
regex I can remember myself having was

1. mainly to match spaces (as of last commit
   I can do that with backslash escapes), and
2. using the "$" end-of-line anchor (I could
   add that as simple style dispatcher later).