.emacs: fix/workaround show-paren-mode quirk
For some reason the function `blink-matching-open`
is called in the middle of Vertico/Orderless
minibuffer completion, which causes a huge UX
freeze when searching for a string that ends in
a closing parenthesis.

Entirely disabling that function doesn't seem to
actually break or even change anything that I've
noticed so far, so I'm gonna try this for a while.