.emacs: datetime-parse: just use first n integers
For example,

    1 2 3 4 5 6 7

will now parse as

    2001-02-03T04:05:06

and not as

    2002-03-04T05:06:07

.

The original motivation for using the last
n integers was to allow a non-backtracking
typing experience for corrections, but in
practice this works poorly, since you can't
make a point correction in this approach by
appending an integer: all integer bindings
shift by one, which is never what we'd want.
And at the time, I didn't have a generic
capability to nix words in the input string
by retyping them with "x" prefixed.

(Sadly I did not bother pulling the creation
of datetime-parse into tiny atomic commits,
so the code to use the last instead of first
integers is already in the initial commit of
this feature.)

Switching back to the naive way of just using
the first n integers both strictly simplifies
and shortens the code, and makes for a more
intuitive UX.