.emacs: read-key-sequence-in-keymap: fix signature
Since the "prompt" argument is mandatory for
both the built-in `read-key-sequence` and my
`read-key-sequence-in-keymap`, make the
signature of the latter actually reflect that.

This was one of those "not sure what's best,
gonna try it the quick and dirty way and figure
it out by seeing which issue becomes the problem
first." And now the first nuisance came from not
having that information in the signature, since
I couldn't remember if that argument mandatory.