.emacs: advice-list function
Given how heavily I use the advice system,
a few times now I've wanted to check if a
function has advice on it. `advice-mapc`
is pretty annoying to manually call it;
and in the worst case, `describe-function`
has a whole lot of text with advice
information buried in the middle (and in
principle the advice information in the
help text could even be false, since the
adding of advice mutates the help string
to add that text, which could get messed
up by other mutations of the doc string).