.emacs: reorder code lines
Moving `with-hook` up with my other `with-*`
macros.

Originally put it with `add-single-use-hook`
but it could really go either way, intuitively
I sorta expect all of my `with-` macros to be
near each other, and the extra tipping impetus
(but not necessarily an extra necessary reason)
is that this moves its definition before its
use in `command-execute-in-keymap`.