.emacs: consult-history-execute: fix Enter key
Long ago, I did not know how to express "do
whatever is bound on Enter" in Emacs lisp,
and so I wrote `consult-history-execute` to
queue up an Enter keypress as if it was user
input immediately after it ran.

This worked wonderfully, except that

1. repeating the command created two undo
   steps instead of one (but this didn't
   really bother me), and

2. when sending a command to another Emacs
   pane with window-state-send-operator,
   the Enter would be interpreted as having
   been input by the user after focus had
   already returned from the window to which
   the `consult-history-execute` was sent.

Now that I know how to do key lookups better,
this is absolutely trivial to fix.