.emacs: integrate undo-tree modes & Evil state UX
Undo Tree's visualizer now gets different Evil
states for its normal and "selection" modes,
with vi-style keybindings for entering/exiting
the "selection" mode, and the benefit of my
distinct vi state color-coding.

Undo Tree's visualizer starts in Evil's "replace"
state, since the "main" keypresses in it
actually replace buffer contents (this is a
bit of a stretch, but it's mnemonic enough).

I have "replace" state color-coded yellow, so
it has a nice distict visual association with
mutation that replaces buffer contents (but in
a way that could still be taken back with the
undo functionality, so it really fits here).

Undo Tree's visualizer's "selection" mode
remains linked to Evil's "motion" state. It
is most semantically fitting since the main
difference/purpose of the "selection" mode
is to just move around the undo tree without
changing the underlying buffer.

To enter Undo Tree's visualizer's "selection"
mode, the keybind is "i", because:

1. The closest vi analog is going from normal
   state into insert state (and specifically
   without an accompanying cursor movement,
   hence not "a", "I", or "A") - I have strong
   muscle memory with this, to some extent I
   associate "insert" state with *passthrough*
   (keys just go through to the underlying
   file when editing, or to the underlying
   program when using Eshell+Eat, and so on),
   and "insert" state is the *primary* (by far
   most commonly used) state which is entered
   from and escapes to vi's "normal" state (so
   there's a very strong parallel between vi's
   normal:insert states and these Undo Tree
   replace:motion states).

2. The good alternatives are "s" (for the term
   used by Undo Tree itself: "selection") and
   "m" (for "motion"), but

   * those don't have everythibg in the first
     point going for them,

   * "s" is taken by my "split/window" state,
     which is far too valuable to me to lose,
     and

   * "m" gets used a lot in Emacs for "mark",
     and in vi for markers, and I want to keep
     that key open for some other undo tree
     functionality which is more analogous to
     those two things.

I have "motion" state color-coded cyan, so
it has great contrast with the yellow of
"replace" state, and there is even a nice
semantic/mnemonic parallel between the above
concepts and how Undo Tree gets yellow for
"replace" and cyan for "motion" while most of
my other vi-style buffers get red for "normal"
and green for "insert".

Feeling very happy and satisfied about this
one. Really came together into a functional
and aesthetically pleasing whole.