.emacs: Undo Tree + Evil: bind window state s/S
This had to be done in a separate section (or
the whole undo-tree+evil section has to move
down just for these two keybinds) since it has
the extra dependency on my split/window state
(and thus on ace-window to implement that).

(For the most part I try to keep depencencies
explicit, obvious, and narrowly declared.
Since the rest of the undo-tree+evil config
doesn't use ace-window or window-state code,
having it inside a section with that as a
dependency would be misleading.)