.emacs: Undo Tree + Evil: fix overwrite-mode suppression
Somehow the mouse down event (which calls
`mouse-drag-region`) while in Evil's replace
state causes overwrite-mode to get flipped
back on (the hack of putting the Undo Tree
buffer into "replace" state relies on turning
that off).

I suspect the handling of click+drag somehow
round-trips through Evil's visual state and
the return to replace state is what turns on
overwrite-mode again.

This fix is starting to pile hack on top of
hack, and eventually it would be better to
define a custom Evil state or expand my
existing code for color-coding the modeline
to Evil state to enable code to override the
color temporarily in a buffer.