.emacs: diff-unsaved-changes: pass file name
This chance makes the underlying diff command see
the actual name of the file instead of "file",
and the actual name of the file with "unsaved "
prefixed instead of just the name "unsaved".

The main benefit is that diff-enhancers like Git
Delta are able to take advantage of the extension
in the file name to add syntax highlighting to
diffs (and I presume this includes Difftastic but
I haven't looked into how that implements language
detection).

A negligible additional benefit is that Git Delta's
output now shows for example "foo.c -> unsaved foo.c"
instead of "file -> unsaved".