gp: use file name in diff
Use the first argument's file name as the
instead of "content".

The biggest benefit is that diff enchancers
like Git Delta will be able to use the file
name's extention to turn on syntax coloring.

A lesser benefit is the nice UX finishing
touch that you see the actual file name in
the output.