.emacs: partial-save + partial-revert: pass file name
This makes my underlying "gp" script see
the actual name of the file instead of
"file". Provided that "gp" also passes
the file name to the underlying Git
commands (and I just updated it so that
it does), this has the same benefits as
the ones described in the last commit:
the UX polish of showing an appropriate
string for the file name, and Git Delta
and so on can detect the language and
add syntax coloring.