.emacs: diff-buffer: handle compressed files
If a buffer is of a compressed file (for example,
a .gz file), Emacs transparently decompresses it
on read and compresses it on write.

For the purposes of taking a buffer diff, this
needs to be disabled when writing the buffer to
a temporary file, so that the diff sees the raw
content.

There is a similar mechanism for files encrypted
with GPG, and we could disable those translations
too, by just nilling `file-name-handler-alist`,
but I would rather hold off on the possibility of
locally leaking the contents of encrypted files
into a temp file (even though it's probably just
in tmpfs, the risk is only from local processes
running as root or the current user, and even
though I don't even know how much I'd be using
the GPG stuff through Emacs).