.emacs: histdir-history--add-newest: simplify conditional
After the last commit, these two nested `when`
forms can become one, and since it's fine if
this branch executes when both `older-duplicate`
is nil *and* there's no last entry (which should
only happen when the whole thing is empty), I'm
just collapsing it to just the inner predicate.

Arguably this is worse for understanding and I
will probably be annoyed with myself for making
this implicit instead of explicit in the code
when I next have to understand it again. But I
want to see how strongly and soon that'll happen.