NOTE: .emacs: histdir-add: pure lisp draft
Start of a pure-lisp implementation of a
`histdir-add` which doesn't rely on an
external `histdir` command:

    (defun histdir-add (entry &optional deduplicate)
        (let ((hash      (secure-hash 'sha256 entry))
              (datetime  (format-time-string "%Y%m%dT%H%M%S,%NZ" nil t))
              (directory (expand-file-name histdir)))
            (let ((string-directory (concat directory "/string/"))
                  (call-directory   (concat directory "/call/")))
                (make-directory string-directory t)
                (make-directory call-directory t)
                (let ((string-file (concat string-directory hash))
                      (call-file   (concat call-directory   datetime)))
                    (with-temp-file string-file
                        (insert entry "\n"))
                    (with-temp-file call-file
                        (insert hash "\n")))
                (when deduplicate
                    ???))))

The deduplication is the hard part, since
it needs to iterate through all files in
`call-directory`, which can be huge, and
we wouldn't want to block Emacs for it.

The one advantage that this approach has
for deduplication is that it would have
access to the in-memory state of the
history, which could know of every call
entry (in a complete implementation), so
it could delete all files whose names
match old call datetimes. This could in
principle miss some duplicates (if they
are added and the implementation misses
that they were added, either due to not
using file system watching or due to
concurrency; or if the implementation
has forgotten about them, which implies
that such an implementation in-memory
can't optimize by only remembering data
for the newest duplicate).