.emacs: evil-ex-nosearch on ":nosearch"/":nos"
Something I've wanted for a really long time
is a way to clear the last search in a way
that stops the "n"/"N" keys from resuming it
(instead of ":nohlsearch"/":noh" which only
hides the highlighting of the search).

In practice, I work around this by typing
something like "/asdfksadf" and this is a
nearly-instant way of achieving the same
thing, but we'll see if ":nos" works just
as well or better (the need to hit the
Shift modifier might make it inferior on a
full/laptop keyboard for me, but on touch
screen keyboards on phones "/" and ":"
have comparable ergonomics and just four
keys to reliably clear search (rather than
a dummy search with unlikely but possible
matches) might be better).

And of course I can always end up binding
`evil-ex-nosearch` to something with better
ergonomics later.