.emacs: add `(provide 'pop-to-command)`
In the long run I'd like to pull some of these
things into their own packages, but I'm putting
that off until I have the time and energy and
drive for polishing and publishing this code.

In the meantime, I want to at least start
explicating the dependencies between my code
and config - putting the higher-level stuff
into `use-package` (or my `use-packages`)
forms which refer to my lower-level "packages".

Starting with `pop-to-command` because that's
the dependency of what I'm currently working
on (see commit soon after this one changing
undo-tree's diff to be consistent with my
other diff stuff).