.emacs: use my diff in undo-tree as well
This makes undo-tree use my `git diff`-based
diffing in a pop-to-command Eshell buffer
instead of its own native diff.

Besides the visual/UX consistency, this also
gives me some UX improvements: for example,
I get annoyed that hitting "q" in Undo Tree's
diff hides that buffer but keeps the extra
split window, which is usually tiny and ends
up redundantly switching back to the buffer
I already have open in another larger window
(because that's the buffer for which I opened
undo tree). A bigger benefit is that the
diff opens as full-sized split window, not a
tiny little one, which actually ends up being
more useful for me.

I think I also like how each new diff appends
to the buffer (because that's how I have my
pop-to-command stuff working) instead of
erasing the old diffs. Personally I find this
easier since it lets me cross-check to prior
diffs as I navigate around. However, if I end
up disliking this, it'll be solvable the same
way as it's solvable for all pop-to-command
uses: I'd just change that function to erase
the buffer on each reuse, possible with a
config variable to control that.