.emacs: make undo-tree window not "dedicated"
The only difference I see from making it not
be a ("soft") dedicated window is that it
shares space with other window splits as I'd
expect - doesn't refuse/resist rebalancing.

In particular, this change makes it so that
any pop-to-buffer call just behaves like a
normal pop-to-buffer - creating a window split
only if none exist and reusing the LRU one
otherwise. (Instead of the default behavior,
which caused any `pop-to-buffer` to split the
remaining half of the screen if undo tree was
already occupying one half.)