.emacs: refactor pop-to-command
Pull all the pop-to-command stuff together,
the way it would be as a package.

For now I'm placing it after Eat and before
Evil in my config, since Evil is still the
huge complected section which contains all
sorts of code - in particular, code which
uses pop-to-command.

In the future, I should be able to pull all
the pop-to-command uses in Evil's section
down, either into a `(use-packages evil
pop-to-command ...)` form, and maybe the
actual functions into yet another "package"
(for the funtions that do all the diffing,
for example) and then this pop-to-command
"package" can move down even further in my
config, assuming it's still in my config
and not yet pulled out into a real package.

In general, I think I'm settling on a
heuristic of putting these config-inlined
"package" definitions just before the first
`use-package` clause that uses them, since
that's the closest to when Emacs would load
and evaluate their definitions if they were
actual external Lisp files.