.emacs: tiny refactor/cleanup
Made the hook which initializes the
minibuffer with a few vi-style keybinds
for Vertico and the right quit function
for my history execute/remove/change
functions a named function, so that the
definition can be updated and evaluated
manually while running without having
to remove the old lambda from the hook.

This whole hook is a bad amateur hack
anyway - just something that Works(tm),
which I cobbled together when I had
very little Emacs programming experience.