.emacs.el: format `interactive` forms idiomatically
A long long time ago, from a position of
relative ignorance, I decided that `(interactive)`
should not be visually laid out the same way as
normal forms that are executed as part of the body.

To some extent I still think this is the Correct
perspective. After all, an interactive declaration
is qualitatively different from a body form and
similar to the argument list: it describes part of
the true complete signature of the function, in a
sense.

But sometime last year as I started coding Emacs
lisp basically full-time, I came to appreciate the
angle from which it makes sense to format functions
so that `(interactive ...)` is visually "in" the
"body" of the function. After all, it can describe
logic which runs whenever the function is invoked
interactively, and in that sense is more like the
other body forms than like an argument list.

So for a long time I've been intending to switch
to the standard way, I just hadn't gotten around
to it, and I was staying consistent within this
file (except where my 80 character limit forced
line wrapping anyway).

I think now if I was going to buck the standard
idiom, I would maybe argue that `(interactive)`
should on its own line but indented extra, like
I do with argument lists that are too long to fit
on the same line as the function name. But I don't
feel strongly enough about that, especially due to
the similarity to body forms making a reasonable
case for keeping it at the same indent level.