.emacs.el: histdir: revert mistaken cleanup
This reverts commit 77e470bcca2100833f412aaf1f459a681017a5f0.
This reverts commit 7bd2e3bd384f4592351db2870a66024ed1ed6b47.

I remembered why I needed to nil out old
history nodes! It was indeed a concurrent
use-case: history cycling with up and
down arrow keys means that there might be
a reference to any history entry for an
arbitrarily long amount of time. Setting
the value to nil ensures that even if that
node is the one the user last cycled to
with Up/Down arrows, the next Up/Down will
be able to keep iterating next/previous
links until it reaches the end or one that
doesn't have a nil value.