.emacs.el: Git rebase with number-prefixed " vC"
Since interactive rebase is basically a more
general superset of commit amend, and "I want
to amend (in any way change) the last N commits"
is a rebase command, it makes sense to expand
the same keybind to handle both, and to make it
switch to rebase based on number. So " 2vC"
will do `git rebase --interactive HEAD~2`.

(Prefixing 1 will do a rebase instead of a
commit amend because in a rebase of one commit
we can still do "drop", so this creates a way
to do "hard" resets of HEAD. Not that I foresee
doing that much, but makes sense to err on the
side of flexibility in this case.)

(This commit was created using both the amend
and rebase branches of the code in this commit!
:D )