.emacs.el: stylistic regex tweak
Personally I think I find "---+" easier
to read than "-\\{3,\\}", because of
the reduced backslash escaping noise.

I think I only did the latter way because
I was trying to stay consistent with the
style Denote uses internally and priority
was on moving fast and not getting caught
up in my traps of thoroughly considering
and optimizing every tiny decision.