.emacs.el: denoted-title-get: fix for unsaved notes
Make `denoted-title-get` called on a not-yet-saved
note file retrieve the title from the front matter
(in memory, as it appears in the unsaved buffer)
instead of prematurely falling back to the title
slug in the path itself.

The bug happens because `denote-file-is-note-p`
needs the file to exist and be a regular file,
which *is* a good check in addition to the file
name check, but a better check might be: that,
*or* if the file is being visited by a buffer.
I don't feel like implementing that better
check right now.

I originally picked `denote-file-is-note-p` for
the semantics: the idea of it is "this is a note
(however that's determined, presumably flexible
to change as better heuristics are found). But
in practice, `denote-filename-is-note-p` ends
up being less wrong here, given the current
implementations.