.emacs.el: denote-file-or-visiting-buffer-is-note-p
Add the function implementing the "better check"
mentioned in the last commit: *either* the file
is a regular file on the file system, *or* there
is a buffer visiting it as a file.

(Note that `find-buffer-visiting` does *not*
return Dired buffers, because they technically
are not "visiting" their directory the same
way that file buffers are visiting their file.
In particular, `buffer-file-name` is set in
buffers visiting files, but is nil in Dired.)