.emacs.el: make " D" unable to delete directories
I haven't used this since I added it, and it's a
bit too dangerous for my tastes now that I've had
more time to think about it (particularly just
now as I had my home directory open in Dired for
quite some time).

At first I implemented a couple extra prompts to
protect the home directory (which really should
then become a configurable list of paths to not
delete), where the first prompt needed a "yes"
answer and the second needed a "no" answer to
protect against holding down the key too long or
mashing "y" a few times if I ever developed that
habit, but then I decided it's just not worth
keeping since I'm not using it.

Another idea would be to require a prefix
argument to allow deleting a directory, or
something else like that. Could also require the
directory delete to answer "r" to some prompt,
(mnemonic: recursive) if I ever want to bring it
back.