.emacs.el: Denoted raw rename on "gC"
This finishes renaming coverage: I can now
rename any and all files with my renaming
keybinds, while getting all the benefits of
`denote-rename-file` for all files and
front matter integration for Denote notes.

In particular, this covers changing the
extension and renames that don't conform
to the Denote naming convention, but if
the file looks like a Denote note after
the rename, its frontmatter will still be
updated to match the new name, and if the
file is a note before the rename, and the
rename didn't change the title in the
file name, the title from the frontmatter
is preserved.