.emacs.el: make read-multiple-choice like y-or-n-p
I always disliked the inconsistency between the
y-or-n prompts and the other short one-letter
prompts. Turns out this is because `y-or-n-p`
has the decency of using `read-from-minibuffer`
instead of a raw `read-event`.

So this commit adds a minimal rewrite of
`read-multiple-choice`. I am sure I will
eventually run into problems with it.

I don't have the stamina for rewriting
`map-y-or-n-p` at this time, but that is
another offender.

The benefit of this rewrite is that a call
to `read-event` locks up the entire UI,
while `read-from-minibuffer` doesn't, which
means it's possible to leave a prompt hanging
and switch to other buffers/windows to do
things. I also get more consistent visuals
this way.

The downside is that basically no code will
expect `read-multiple-choice` to behave this
way, it might lead to misbehaviors too subtle
for me to notice, no buffer calling this
function expects other interactions until the
function returns (casually breaking invariants
this way without even thinking about it is now
possible with this change) and it's not a full
reimplementation (I skipped the help feature
(which I don't use and have never yet wanted)
and the long-form completing-read feature). Yet
another downside is that I don't handle the
non-interactive case, and maybe I'd want that
for testing or something? (Ditto the dialog
box case, but I doubt I'll ever want that).